requirements.txt
pytest~=8.3.2
selene>=2.0.b4
allure-pytest==2.13.5
